Profile Interface with Random Avatar
This PR adds a profile interface where users can choose from 8 random avatars as a temporary solution. The random avatars are available for selection, and this feature will be replaced with a more permanent avatar selection method in the future.
Features created/files changed:

Implemented profile interface with 8 random avatars for selection.
Added functionality to choose an avatar from the 8 available options.
Random avatar selection is temporary and will be updated nextweek.